src/requirements.txt
# Core
numpy==1.26.4
pandas==2.1.3
streamlit==1.28.0
reportlab
# ML / NLP
scikit-learn
sentence-transformers==2.2.2
transformers==4.35.0
faiss-cpu==1.7.4
torch==2.2.0

# APIs / Utilities
google-generativeai==0.3.1
prometheus-client==0.19.0
fpdf==1.7.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
tqdm==4.66.1
python-dotenv==1.0.0
